% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{articlesInfo}
\alias{articlesInfo}
\title{Functional studies}
\usage{
articlesInfo(object, variant.cor, bbdd)
}
\value{
If the variants has undergone some of the following  multifactorial studies or functional assays, it returns the information: Parsons et al.(2019); Lyra et al. (2021); Jia et al.(2021), Drost et al. (2018).
}
\description{
articlesInfo() retrieves information from different multifactorial studies or functional assays
}
\references{
Parsons, M. T., Tudini, E., Li, H., Hahnen, E., Wappenschmidt, B., Feliubadaló, L., ... & Pohl‐Rescigno, E. (2019). Large scale multifactorial likelihood quantitative analysis of BRCA1 and BRCA2 variants: An ENIGMA resource to support clinical variant classification. Human mutation, 40(9), 1557-1578.
Lyra, P. C., Nepomuceno, T. C., de Souza, M. L., Machado, G. F., Veloso, M. F., Henriques, T. B., ... & Monteiro, A. N. (2021). Integration of functional assay data results provides strong evidence for classification of hundreds of BRCA1 variants of uncertain significance. Genetics in Medicine, 23(2), 306-315.
Jia, X., Burugula, B. B., Chen, V., Lemons, R. M., Jayakody, S., Maksutova, M., & Kitzman, J. O. (2021). Massively parallel functional testing of MSH2 missense variants conferring Lynch syndrome risk. The American Journal of Human Genetics, 108(1), 163-175.
Drost, M., Tiersma, Y., Thompson, B. A., Frederiksen, J. H., Keijzers, G., Glubb, D., ... & Tavtigian, S. V. (2019). A functional assay–based procedure to classify mismatch repair gene variants in Lynch syndrome. Genetics in Medicine, 21(7), 1486-1496.
Kato, S., Han, S. Y., Liu, W., Otsuka, K., Shibata, H., Kanamaru, R., & Ishioka, C. (2003). Understanding the function-structure and function-mutation relationships of p53 tumor suppressor protein by high-resolution missense mutation analysis. Proceedings of the National Academy of Sciences of the United States of America, 100(14), 8424–8429. https://doi.org/10.1073/pnas.1431692100
Giacomelli, A. O., Yang, X., Lintner, R. E., McFarland, J. M., Duby, M., Kim, J., Howard, T. P., Takeda, D. Y., Ly, S. H., Kim, E., Gannon, H. S., Hurhula, B., Sharpe, T., Goodale, A., Fritchman, B., Steelman, S., Vazquez, F., Tsherniak, A., Aguirre, A. J., Doench, J. G., … Hahn, W. C. (2018). Mutational processes shape the landscape of TP53 mutations in human cancer. Nature genetics, 50(10), 1381–1387. https://doi.org/10.1038/s41588-018-0204-y
Kotler, E., Segal, E., & Oren, M. (2018). Functional characterization of the p53 "mutome". Molecular & cellular oncology, 5(6), e1511207. https://doi.org/10.1080/23723556.2018.1511207
}
\author{
Elisabet Munté Roca
}
