% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{BLOSUM62}
\alias{BLOSUM62}
\title{BLOSUM62}
\description{
BLOSUM62
}
\references{
K. Malde, The effect of sequence quality on sequence alignment, Bioinformatics, Feb 23, 2008. \url{https://academic.oup.com/bioinformatics/article/24/7/897/296249}
}
\author{
H. Pags and P. Aboyoun
}
