% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{clinVarIds}
\alias{clinVarIds}
\title{clinVarIds variant}
\usage{
clinVarIds(object)
}
\arguments{
\item{object}{obtained with the function varDetails()}
}
\value{
A list containing ClinVar id of the variant of interest and for missense variants it also returns the variants located at the same codon.
}
\description{
clinVarIds variant
}
\references{
https://clinicaltables.nlm.nih.gov/apidoc/variants/v4/doc.html
}
\author{
Elisabet Munté Roca
}
