% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{clinVarInfo}
\alias{clinVarInfo}
\title{ClinVar info}
\usage{
clinVarInfo(clinvar.id, object)
}
\arguments{
\item{clinvar.id}{a list containing clinvarIDs, obtained with function clinVarIds()}
}
\value{
All the information contained in Clinvar for this variant
}
\description{
ClinVar info
}
\references{
Landrum MJ, Lee JM, Benson M, Brown GR, Chao C, Chitipiralla S, Gu B, Hart J, Hoffman D, Jang W, Karapetyan K, Katz K, Liu C, Maddipatla Z, Malheiro A, McDaniel K, Ovetsky M, Riley G, Zhou G, Holmes JB, Kattman BL, Maglott DR. ClinVar: improving access to variant interpretations and supporting evidence. Nucleic Acids Res . 2018 Jan 4. PubMed PMID: 29165669 .
}
\author{
Elisabet Munté Roca
}
