% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{correctHgvsMutalyzer}
\alias{correctHgvsMutalyzer}
\title{Correct variant hgvs nomenclature}
\usage{
correctHgvsMutalyzer(NM, NC = NULL, gene, variant)
}
\arguments{
\item{NM}{Accession number of the transcrit and mRNA from RefSeq}

\item{NC}{Accession number of the chromosome RefSeq. It can be ommited if the variant is exonic.}

\item{gene}{Your gene of interest}

\item{variant}{Your cdna variant of interest}
}
\value{
The correct hgvs nomenclature of your variant from Mutalyzer. If the variant is intronic,
Mutalyzer V3 is checked where as if the variant is not intronic Mutalyzer v2. A list is given, where the variant and the error message can be checked.
}
\description{
Correct variant hgvs nomenclature
}
\examples{
correctHgvsMutalyzer(NM="NM_000051.3", gene="ATM", variant="c.3436G>A")
correctHgvsMutalyzer(NM="NM_007294.3", NC="NC_000017.10", gene="BRCA1", variant="c.-19-85C>G")
}
\references{
"Lefter M et al. (2021). Mutalyzer 2: Next Generation HGVS Nomenclature Checker. Bioinformatics, btab051" (direct link)
}
\author{
Elisabet Munté Roca
}
