% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{domainProt}
\alias{domainProt}
\title{Variant's domain from UCSC}
\usage{
domainProt(chr, start, end)
}
\arguments{
\item{chr}{Chromosome where the variant is located}

\item{start}{genomic start of the variant}

\item{end}{genomic end of the variant}
}
\value{
The protein domain where the variant is located (Uniprot)
}
\description{
Variant's domain from UCSC
}
\author{
Elisabet Munté Roca
}
