% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{ex_vaRbatch}
\alias{ex_vaRbatch}
\title{example_input_vaRbatch}
\description{
example_input_vaRbatch
}
\references{
\url{https://github.com/emunte/vaRHC/blob/main/data/ex_vaRbatch.txt}
}
\author{
Elisabet Munté Roca \email{emunte@idibell.cat}
}
\keyword{data}
