% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{flossiesInfo}
\alias{flossiesInfo}
\title{Information from FLOSSIES db}
\usage{
flossiesInfo(nomen.flossies, nomen.gnomAD = NULL, gene)
}
\arguments{
\item{nomen.flossies}{character. FlOSSIES Id of the variant of interest. If you don't know it it can be set to NULL}

\item{nomen.gnomAD}{character. By default is null. If  nomen.flossies is not known you can use gnomAD nomenclature. gnomAD nomenclature can be infered with gnomADnomen() function.}
}
\value{
A dataframe with the information of the variant listen in FLOSSIES db (Fabulous Ladies Over Seventy).
}
\description{
Information from FLOSSIES db
}
\examples{
####Knowing FLOSSIES nomenclature
nomenclature.flossies <- "17-41245700-AGA-"
flossiesInfo (nomen.flossies=nomenclature.flossies)
}
\references{
https://whi.color.com/
}
\author{
Elisabet Munté
}
