% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{gencode_spliceaihg19_}
\alias{gencode_spliceaihg19_}
\title{gencode_spliceai_hg19}
\description{
gencode_spliceai_hg19
}
\references{
adapted from SpliceAI-lookup \url{https://github.com/broadinstitute/SpliceAI-lookup/blob/master/annotations/gencode.v38lift37.annotation.txt.gz}
}
\keyword{data}
