% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{gene_specific}
\alias{gene_specific}
\title{gene_specific}
\description{
gene_specific
}
\references{
\url{https://github.com/emunte/vaRHC/blob/main/data/gene_specific.txt}
}
\author{
Elisabet Munté Roca \email{emunte@idibell.cat}
}
