% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{geneSpecific}
\alias{geneSpecific}
\title{Know if the gene has special ACMG rules}
\usage{
geneSpecific(gene, gene.specific.df)
}
\arguments{
\item{gene}{the gene of interest}

\item{gene.specific.df}{By default is NULL, it uses the default parameters described in README. If you would like to change some defaults or include another gene, a template can be downloaded from Github: https://github.com/emunte/Class_variants/tree/main/documents/gen_especific.csv and some parameters can be modified taking into account your preferences}
}
\value{
The name of the gene if there are special ACMG rules for it and if not it returns general ACMG rules.
if it follows general ACMG rules (only for hereditary cancer related genes)
}
\description{
Know if the gene has special ACMG rules
}
\author{
Elisabet Munté Roca
}
