% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{gnomADcov}
\alias{gnomADcov}
\title{gnomADcov}
\usage{
gnomADcov(
  assembly,
  track,
  gnomAD.nomenclature,
  chr = NULL,
  start = NULL,
  end = NULL,
  mean.value = TRUE
)
}
\arguments{
\item{assembly}{character hg19 or hg38}

\item{track}{character.It can be either Exomes or Genomes. GnomAD}

\item{gnomAD.nomenclature}{character. GnomAD variant nomenclature. If ID is not known it can be set to NA. Chr, start and end must be given then}

\item{chr}{numeric. Chromosome of interest. By default is NULL}

\item{start}{numeric.Start of the region of interest. By default is NULL}

\item{end}{numeric. End of the region of interest. By default is NULL}

\item{mean}{logical. By default is TRUE. If the region is more than 1 bp long it returns the mean coverage or if FALSE it returns the coverage of each position}
}
\value{
Numeric value of the coverage region
}
\description{
gnomADcov
}
\references{
Karczewski, K.J., Francioli, L.C., Tiao, G. et al. The mutational constraint spectrum quantified from variation in 141,456 humans. Nature 581, 434–443 (2020). https://doi.org/10.1038/s41586-020-2308-7
}
\author{
Elisabet Munté Roca
}
