% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{gnomADmerge}
\alias{gnomADmerge}
\title{gnomADmerge}
\usage{
gnomADmerge(gnom, specific.info, bbdd)
}
\arguments{
\item{gnom}{character. Variant in gnomAD nomenclature. Use gnomADnomen() function if you don't have gnomAD nomenclature but yes cdna annotation.}

\item{specific.info}{data obtained with the geneSpecific () function}
}
\value{
variant's information located in gnomAD in the gnomAD v2.1,1 (non-cancer) and gnomAD v2.1.1 (non-neuro). It uses connection to mySQL server.
}
\description{
gnomADmerge
}
\examples{
#####Knowing gnomAD nomenclature
nomenclature <- "17-41258474-T-C"
gene <- "BRCA1"
specific <- geneSpecific(nomenclature, geneSpecific(gene, gene.specific.df))
gnomADmerge(nomenclature, specific)
#####Not knowing gnomAD nomenclature
#' NM <- "NM_007294.3"
NC <- "NC_000017.10"
gene <- "BRCA1"
CCDS <- "CCDS11453.1"
variant <- "c.211A>G"
mutalyzer.info <- correctHgvsMutalyzer(NM, NC,  gene, variant)
variant.info <- varDetails(NM,NC, CCDS, gene, mutalyzer.info, skip.pred=FALSE)
nomenclature <- gnomADnomen(variant.info)
specific <- geneSpecific(nomenclature, geneSpecific(gene, gene.specific.df))
gnomADmerge(nomenclature, specific)
}
\references{
Karczewski, K.J., Francioli, L.C., Tiao, G. et al. The mutational constraint spectrum quantified from variation in 141,456 humans. Nature 581, 434–443 (2020). https://doi.org/10.1038/s41586-020-2308-7
}
\author{
Elisabet Munté Roca
}
