% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rutils.R
\name{readUrl}
\alias{readUrl}
\title{read url}
\usage{
readUrl(description, verbose = FALSE)
}
\arguments{
\item{description}{url to read}

\item{verbose}{logical. By default is FALSE``}
}
\description{
read url
}
