% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_info.R
\name{traPscore}
\alias{traPscore}
\title{TrapScore}
\usage{
traPscore(gnom)
}
\description{
TrapScore
}
\references{
Gelfman, S., Q. Wang, K.M McSweeney, Z. Ren, F. La Carpia, M. Halvorsen, K. Schoch, F. Ratzon, E.L. Heinzen, M.J. Boland, S. Petrovski and D. B. Goldstein. 2017. Annotating Pathogenic Non-Coding Variants in Genic Regions. Accepted June 5th, 2017.
}
