% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors_info.R
\name{ucscRPredictor}
\alias{ucscRPredictor}
\title{UCSC predictors}
\usage{
ucscRPredictor(track, object)
}
\arguments{
\item{track}{can be either phyloP100wayAll, phastCons100way or allHg19RS_BW}

\item{chr}{chromosome of interest}

\item{start}{start genomic position}

\item{end}{end genomic position}
}
\value{
the value of the ucsc predictor
}
\description{
UCSC predictors
}
\author{
Elisabet Munté
}
