% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{vaRbatch}
\alias{vaRbatch}
\title{vaRbatch()}
\usage{
vaRbatch(
  assembly = "hg19",
  all.variants,
  gene.specific.df = NULL,
  browser = "firefox",
  spliceai.program = FALSE,
  spliceai.reference = NULL,
  spliceai.annotation = system.file("docs", "gencode.v38lift37.annotation.txt", package =
    "vaRHC"),
  spliceai.distance = 1000,
  spliceai.masked = 1,
  provean.program = FALSE,
  print.data.frame = TRUE,
  excel.results = FALSE,
  path.copy.file = NULL
)
}
\arguments{
\item{assembly}{character hg19 or hg38}

\item{gene.specific.df}{By default is NULL, it uses the default parameters described in README. If you would like to change some defaults or include another gene, a template can be downloaded from Github: https://github.com/emunte/Class_variants/tree/main/documents/gen_especific.csv or in the package docs folder and some parameters can be modified taking into account your preferences}

\item{excel.results}{Logical. By default is FALSE and no excel file would be produced. If TRUE and excel file will be saved}

\item{path.copy.file}{By default is NULL. If excel.results param is set to TRUE, path.copy.file must provide the path where the excel file must be saved. If not provided it will be saved in the working directory.}

\item{gene}{gene of interest}

\item{variant}{variant of interest in cdna}

\item{NM}{Accession number of the transcrit and mRNA from RefSeq. By default is NULL and vaRHC will consider the ones detailed in README file. Be careful if you use a different NM because the program has not been validated for it. If you provide a different NM,  NC and CCDS must also be provided.}

\item{NC}{Accession number of the chromosome RefSeq. It can be ommited if the variant is exonic. By default is NULL and vaRHC will consider the ones detailed in README file. Be careful if you use a different NM because the program has not been validated for it. If you provide a different NC, NM and CCDS must also be provided.}

\item{CCDS}{Consensus CD id https://www.ncbi.nlm.nih.gov/projects/CCDS/CcdsBrowse.cgi. By default is NULL and vaRHC will consider the ones detailed in README file. Be careful if you use a different CCDS because the program has not been validated for it. If you provide a different CCDS, NM and NC must also be provided. Current version only works for hg19.}
}
\description{
vaRbatch function allows to perform vaR function in batch. It also returns a logfile.
}
\author{
Elisabet Munté Roca
}
