% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{vaRclass}
\alias{vaRclass}
\title{vaRclass}
\usage{
vaRclass(all.information)
}
\arguments{
\item{all.information}{obtained with the getVaRinfo function}
}
\value{
the classification of the variant  and the criteria given following ACMG updated rules and gene specific for BRCA1, BRCA2, PALB2, ATM, MMR
}
\description{
vaRclass
}
\author{
Elisabet Munté Roca
}
