% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{varDetails}
\alias{varDetails}
\title{Variant info from Ensembl VEP}
\usage{
varDetails(
  NM,
  NC = NULL,
  CCDS,
  gene,
  variant,
  variant.mutalyzer = NULL,
  skip.pred = FALSE
)
}
\arguments{
\item{NM}{Accession number of the transcrit and mRNA from RefSeq}

\item{NC}{Accession number of the chromosome RefSeq. It can be ommited if the variant is exonic.}

\item{CCDS}{The Consensus CDS code}

\item{gene}{The gene where the variant is located}

\item{variant.mutalyzer}{By default is null. Obtained with the correctHgvsMutalyzer function contains all information of the variant from mutalyzer.}

\item{skip.pred}{By default is false. If you are predicting the consequence of a skipping then it should be set to TRUE}
}
\value{
The chromosome, the initial genomic coordinate, the final genomic
coordinate, the reference allele, the alternative allele, the most important
variant consequence and if the variant is located in a protein domain
}
\description{
Variant info from Ensembl VEP
}
\examples{
NM <- "NM_007294.3"
NC <- "NC_000017.10"
gene <- "BRCA1"
CCDS <- "CCDS11453.1"
variant <- "c.211A>G"
mutalyzer.info <- correctHgvsMutalyzer(NM, NC,  gene, variant)
varDetails(NM,NC, CCDS, gene, mutalyzer.info, skip.pred=FALSE)
}
\references{
McLaren W, Gil L, Hunt SE, Riat HS, Ritchie GR, Thormann A, Flicek P, Cunningham F. The Ensembl Variant Effect Predictor. Genome Biology Jun 6;17(1):122. (2016) doi:10.1186/s13059-016-0974-4
}
\author{
Elisabet Munté Roca
}
