% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaRHC-package.R
\docType{package}
\name{vaRHC-package}
\alias{vaRHC}
\alias{vaRHC-package}
\title{vaRHC: ACMG variant classification semi-automated}
\description{
ACMG variant classification.
}
\author{
\strong{Maintainer}: Elisabet Munté Roca \email{emunte@idibell.cat} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
