% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{vaRinfo}
\alias{vaRinfo}
\title{vaRinfo ()}
\usage{
vaRinfo(
  assembly,
  gene,
  variant,
  NM = NULL,
  NC = NULL,
  CCDS = NULL,
  gene.specific.df = NULL,
  browser = "firefox",
  spliceai.program = FALSE,
  spliceai.reference = NULL,
  spliceai.annotation = "../docs/gencode.v38lift37.annotation.txt",
  spliceai.distance = 1000,
  spliceai.masked = 1,
  provean.program = FALSE
)
}
\arguments{
\item{assembly}{character hg19 or hg38}

\item{gene}{gene of interest}

\item{variant}{variant of interest in cdna}

\item{NM}{refSeq nomenclature. By default is NULL and vaRHC will consider the ones detailed in README file. Be careful if you use a different NM because the program has not been validated for it. If you provide a different NM,  NC and CCDS must also be provided.}

\item{NC}{refSeq nomenclature. By default is NULL and vaRHC will consider the ones detailed in README file. Be careful if you use a different NM because the program has not been validated for it. If you provide a different NC, NM and CCDS must also be provided.}

\item{CCDS}{Consensus CD id https://www.ncbi.nlm.nih.gov/projects/CCDS/CcdsBrowse.cgi. By default is NULL and vaRHC will consider the ones detailed in README file. Be careful if you use a different CCDS because the program has not been validated for it. If you provide a different CCDS, NM and NC must also be provided. Current version only works for hg19.}

\item{gene.specific.df}{By default is NULL, it uses the default parameters described in README. If you would like to change some defaults or include another gene, a template can be downloaded from Github: https://github.com/emunte/Class_variants/tree/main/documents/gen_especific.csv and some parameters can be modified taking into account your preferences}
}
\value{
information about the variant.
}
\description{
Collect different types of criteria used by ACMG to classify variants.
}
\examples{
vaRinfo("hg19", "MSH6", "c.211A>G", spliceai.program = TRUE, spliceai.reference = "/home/emunte/hg19.fa")
}
\references{
Richards, S., Aziz, N., Bale, S., Bick, D., Das, S., Gastier-Foster, J., Grody, W. W., Hegde, M., Lyon, E., Spector, E., Voelkerding, K., Rehm, H. L., & ACMG Laboratory Quality Assurance Committee (2015). Standards and guidelines for the interpretation of sequence variants: a joint consensus recommendation of the American College of Medical Genetics and Genomics and the Association for Molecular Pathology. Genetics in medicine : official journal of the American College of Medical Genetics, 17(5), 405–424. https://doi.org/10.1038/gim.2015.30
}
\author{
Elisabet Munté Roca
}
