% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rutils.R
\name{vcfToCodingDNA}
\alias{vcfToCodingDNA}
\title{vcfToCodingDNA()}
\usage{
vcfToCodingDNA(
  file.vcf,
  assembly,
  annotation = "LRG",
  log.dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{file.vcf}{A filename for a variant call format (vcf or compressed *.vcf.gz.format) file.}

\item{assembly}{character. It can only be "hg19" or "hg38" assembly.}

\item{annotation}{character.  It can only be LRG or MANE_select. Variants will be annotated considering LRG or MANE_select transcripts.}

\item{log.dir}{path where log folder will be created. By default log folder will be created in working diretory.}

\item{verbose}{logic. By default is FALSE. If TRUE, it includes status messages (if any).}
}
\value{
vcfToHgvs returns a tibble  of four columns (gene, variant, NM, genomic). The number of rows will be equivalent to the number of variants annotated. A log file is generated and stored in log.dir path.
}
\description{
This function allows to convert variant call format (vcf) file to a tibble with the variants annotated in coding DNA nomenclature considering MANE select transcript.
}
\details{
The function \emph{vcfToCodingDNA} allows to annotate variants stored in a variant call format (vcf or compressed vcf) in coding DNA nomenclature.
To do so it uses Mutalyzer v3 API, thus connection to internet is needed.
This function is intended for small vcfs. Vcfs resulting from WES or WGS will take too long (although they work as well).
The output of this function can be used to run \emph{vaRbatch} function.
}
\examples{
\dontrun{
file <- "/path/to/vcf/file.vcf"
variants <- vcfToCodingDNA(file.vcf = file, assembly="hg19")
}
}
\references{
Morales, J., Pujar, S., Loveland, J.E. et al. A joint NCBI and EMBL-EBI transcript set for clinical genomics and research. Nature. 2022 Apr;604(7905):310-315. PubMed; PubMed Central; DOI: 10.1038/s41586-022-04558-8 . https://www.ncbi.nlm.nih.gov/refseq/MANE/
Lefter M et al. (2021). Mutalyzer 2: Next Generation HGVS Nomenclature Checker. Bioinformatics, 2021 Sep 15; 37(28):2811-7 (version 3: https://mutalyzer.nl/ )
Knaus, Brian J., and Niklaus J. Grünwald. "vcfr: a package to manipulate and visualize variant call format data in R." Molecular ecology resources 17.1 (2017): 44-53. (http://dx.doi.org/10.1101/041277)
}
\author{
Elisabet Munté Roca
}
