% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{geneSpecific}
\alias{geneSpecific}
\title{Know if the gene has special ACMG rules}
\usage{
geneSpecific(gene, gene.specific.df)
}
\arguments{
\item{gene}{the gene of interest}

\item{gene.specific.df}{By default is NULL, it uses the default parameters described in README. If you would like to change some defaults or include another gene, a template can be downloaded from Github: https://github.com/emunte/Class_variants/tree/main/documents/gen_especific.csv and some parameters can be modified taking into account your preferences}
}
\value{
The name of the gene if there are special ACMG rules for it and if not it returns general ACMG rules.
if it follows general ACMG rules (only for hereditary cancer related genes)
}
\description{
Know if the gene has special ACMG rules
}
\examples{
#using the default
geneSpecific("BRCA1")
geneSpecific("RAD51C")
geneSpecific("ERCC5")

#changing the csv file
file1 <- file.path("./gene_specific.csv")
gene.specific.df <- read.csv(file1, sep=",", header=TRUE,row.names=1, stringsAsFactors = FALSE)
geneSpecific("BRCA1", gene.specific.df)
geneSpecific("RAD51C", gene.specific.df)
geneSpecific("ERCC5", gene.specific.df)
}
\references{
Richards, S., Aziz, N., Bale, S., Bick, D., Das, S., Gastier-Foster, J., Grody, W. W., Hegde, M., Lyon, E., Spector, E., Voelkerding, K., Rehm, H. L., & ACMG Laboratory Quality Assurance Committee (2015). Standards and guidelines for the interpretation of sequence variants: a joint consensus recommendation of the American College of Medical Genetics and Genomics and the Association for Molecular Pathology. Genetics in medicine : official journal of the American College of Medical Genetics, 17(5), 405–424. https://doi.org/10.1038/gim.2015.30
Tavtigian, S. V., Harrison, S. M., Boucher, K. M., & Biesecker, L. G. (2020). Fitting a naturally scaled point system to the ACMG/AMP variant classification guidelines. Human mutation, 41(10), 1734–1737. https://doi.org/10.1002/humu.24088
Garrett, A., Durkie, M., Callaway, A., Burghel, G. J., Robinson, R., Drummond, J., Torr, B., Cubuk, C., Berry, I. R., Wallace, A. J., Ellard, S., Eccles, D. M., Tischkowitz, M., Hanson, H., Turnbull, C., & CanVIG-UK (2021). Combining evidence for and against pathogenicity for variants in cancer susceptibility genes: CanVIG-UK consensus recommendations. Journal of medical genetics, 58(5), 297–304. https://doi.org/10.1136/jmedgenet-2020-107248
ClinGen CDH1 Expert Panel Specifications to the ACMG/AMP Variant Interpretation Guidelines Version 3: https://clinicalgenome.org/site/assets/files/7118/clingen_cdh1_acmg_specifications_v3.pdf
ClinGen PTEN Expert Panel Specifications to the ACMG/AMP Variant Interpretation Guidelines Version 2: https://clinicalgenome.org/site/assets/files/4000/clingen_pten_acmg_specifications_v2.pdf
ClinGen InSiGHT Hereditary Colorectal Cancer/Polyposis Variant Curation Expert Panel Specifications to the ACMG/AMP Variant Interpretation Guidelines Version 1 (draft): https://www.insight-group.org/content/uploads/2021/11/DRAFT_Nov_2021_TEMPLATE_SVI.ACMG_Specifications_InSiGHT_MMR_V1.pdf
}
\author{
Elisabet Munté Roca
}
