% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{NMparam}
\alias{NMparam}
\title{Get transcript information}
\usage{
NMparam(gene, NM = NULL, CCDS = NULL)
}
\arguments{
\item{gene}{Your gene of interest
If the gene doesn't exist in the table, an error message is obtained}

\item{NM}{Accession number of the transcrit and mRNA from RefSeq.}

\item{CCDS}{Consensus CDS id https://www.ncbi.nlm.nih.gov/projects/CCDS/CcdsBrowse.cgi.}
}
\value{
The NM, NC and CCDS for that gene
}
\description{
Get transcript information
}
\author{
Elisabet Munté Roca
}
