% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{NMparam}
\alias{NMparam}
\title{Get transcript information}
\usage{
NMparam(gene, NM = NULL, NC = NULL, CCDS = NULL)
}
\arguments{
\item{gene}{Your gene of interest
If the gene doesn't exist in the table, an error message is obtained}

\item{nm.info}{By default NULL. If not a valid NM (Ref-seq Accession number) must be provided. Be careful because only default NM have been tested.}
}
\value{
The NM, NC and CCDS for that gene
}
\description{
Get transcript information
}
\examples{

NMparam(gene="BRCA1")
NMparam(gene="TP53")
NMparam(gene="TP53", NM = "NM_001126113.2", NC = "NC_000017.10", CCDS= "CCDS45606")
}
\author{
Elisabet Munté Roca
}
