% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaRHC-package.R
\docType{package}
\name{vaRHC-package}
\alias{vaRHC}
\alias{vaRHC-package}
\title{vaRHC: Variant Classification in Hereditary Cancer Following Specific ACMG guidelines}
\description{
automate, as far as possible, the variant classification process for hereditary cancer genes following specific ACMG guidelines
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emunte/vaRHC}
  \item Report bugs at \url{https://github.com/emunte/vaRHC/issues}
}

}
\author{
\strong{Maintainer}: Elisabet Munté Roca \email{emunte@idibell.cat} (\href{https://orcid.org/0000-0001-8396-8284}{ORCID})

}
\keyword{internal}
