% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{vaRinfo}
\alias{vaRinfo}
\title{vaRinfo ()}
\usage{
vaRinfo(
  assembly,
  gene,
  variant,
  NM = NULL,
  NC = NULL,
  CCDS = NULL,
  gene.specific.df = NULL,
  browser = "firefox",
  spliceai.program = FALSE,
  spliceai.reference = NULL,
  spliceai.annotation = system.file("data", "gencode_spliceai_hg19.txt", package =
    "vaRHC"),
  spliceai.distance = 1000,
  spliceai.masked = 1,
  provean.program = FALSE,
  provean.sh = NULL
)
}
\arguments{
\item{assembly}{character hg19 or hg38}

\item{gene}{gene of interest}

\item{variant}{variant of interest in cdna}

\item{NM}{refSeq nomenclature. By default is NULL and vaRHC will consider the ones detailed in README file. Be careful if you use a different NM because the program has not been validated for it. If you provide a different NM,  NC and CCDS must also be provided.}

\item{NC}{refSeq nomenclature. By default is NULL and vaRHC will consider the ones detailed in README file. Be careful if you use a different NM because the program has not been validated for it. If you provide a different NC, NM and CCDS must also be provided.}

\item{CCDS}{Consensus CD id https://www.ncbi.nlm.nih.gov/projects/CCDS/CcdsBrowse.cgi. By default is NULL and vaRHC will consider the ones detailed in README file. Be careful if you use a different CCDS because the program has not been validated for it. If you provide a different CCDS, NM and NC must also be provided. Current version only works for hg19.}

\item{gene.specific.df}{By default is NULL, it uses the default parameters described in README. If you would like to change some defaults or include another gene, a template can be downloaded from Github: https://github.com/emunte/Class_variants/tree/main/documents/gen_especific.csv and some parameters can be modified taking into account your preferences}

\item{browser}{Which browser to start Rselenium server. By default is "firefox" (the recommended). If you do not have firefox installed try either "chrome" or "phantomjs".}

\item{spliceai.program}{Logical. By default is FALSE and it is assumed that SpliceAI program is not installed in your computer. If this parameter is FALSE, the program will only classify substitutions and simple deletion variants taking into account a spliceAI distance of 1000 and will show masked results. If you want to classify other variants please install SpliceAI (https://pypi.org/project/spliceai/) and set to TRUE the parameter.}

\item{spliceai.reference}{Path to the Reference genome hg19 fasta file. Can be downloaded from http://hgdownload.cse.ucsc.edu/goldenPath/hg19/bigZips/hg19.fa.gz . By default is NULL and it will only be taken into account if spliceai.program is set to TRUE.}

\item{spliceai.annotation}{Path to gene annotation file. By default it uses the file stored in extdata folder: "../extdata/gencode_spliceai_hg19.txt"}

\item{spliceai.distance}{Integer. Maximum distance between the variant and gained/lost splice site (default: 1000)}

\item{spliceai.masked}{Mask scores representing annotated acceptor/donor gain and unannotated acceptor/donor loss (default: 1)}

\item{provean.program}{Logical. By default is FALSE and it is assumed that provean program is not installed in your computer.}

\item{provean.sh}{Path to provean.sh. By default is NULL and will only be taken into account if provean.program is set to TRUE.}
}
\value{
information about the variant.
}
\description{
Collect different types of criteria used by ACMG to classify variants.
}
\examples{
vaRinfo("hg19", "MLH1", "c.1AG>")
vaRinfo("hg19", "MSH6", "c.211A>G", spliceai.program = TRUE, spliceai.reference = "./hg19.fa")
}
\references{
Richards, S., Aziz, N., Bale, S., Bick, D., Das, S., Gastier-Foster, J., Grody, W. W., Hegde, M., Lyon, E., Spector, E., Voelkerding, K., Rehm, H. L., & ACMG Laboratory Quality Assurance Committee (2015). Standards and guidelines for the interpretation of sequence variants: a joint consensus recommendation of the American College of Medical Genetics and Genomics and the Association for Molecular Pathology. Genetics in medicine : official journal of the American College of Medical Genetics, 17, 405–424. https://doi.org/10.1038/gim.2015.30
}
\author{
Elisabet Munté Roca
}
